Split PDP partials into their own files.
